Refactor `TimeUnit` to support additional units (HOURS, DAYS, WEEKS, MONTHS, YEARS) and update calculations.
- Move `TimeUnit` to lib package
