+++ code/lib/src/main/java/org/betonquest/betonquest/lib/argument/type/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Argument type classes.
+ */
+package org.betonquest.betonquest.lib.argument.type;
